requirements.txt
# Requirements otimizado para Azure App Service
# Usa versões pré-compiladas para evitar build pesado

fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# Core runtimes
requests==2.31.0
networkx==3.2.1
openai==1.52.0
httpx==0.27.0
PyJWT==2.8.0
reportlab==4.0.7
unidecode==1.3.7
websockets==12.0
python-multipart==0.0.6